HW4 Task1(return value code reformatted, raise exception added) and Task1 test (check for exceptions, file cleanup added)

+++ tests/homework4/test_task1.py
@@ -1,6 +1,8 @@
 ﻿import os
 import tempfile
 
+import pytest
+
 from homework4.task1 import read_magic_number
 
 
@@ -10,7 +12,8 @@ def test_read_magic_number_pos():
     fp.write('2')
     fp.close()
     result = read_magic_number(fp.name)
-    os.remove(fp.name)
+    if os.path.exists(fp.name):
+        os.remove(fp.name)
     assert result
 
 
@@ -20,7 +23,8 @@ def test_read_magic_number_neg():
     fp.write('-1')
     fp.close()
     result = read_magic_number(fp.name)
-    os.remove(fp.name)
+    if os.path.exists(fp.name):
+        os.remove(fp.name)
     assert not result
 
 
@@ -29,12 +33,13 @@ def test_read_magic_number_not_number():
     fp = tempfile.NamedTemporaryFile(mode='w+t', delete=False)
     fp.write('ddd')
     fp.close()
-    result = read_magic_number(fp.name)
-    os.remove(fp.name)
-    assert not result
+    with pytest.raises(ValueError):
+        read_magic_number(fp.name)
+        if os.path.exists(fp.name):
+            os.remove(fp.name)
 
 
 def test_read_magic_number_nofile():
     """Non-existing file"""
-    result = read_magic_number('1233')
-    assert not result
+    with pytest.raises(ValueError):
+        read_magic_number('1233')
